Lannootree_processor/processorAPI/requirements.txt
imageio==2.22.4
opencv-python==4.6.0.66
paho_mqtt==1.6.1
redis==4.3.5
